re-organized code : subfolders contact and lowpassfilter + documented new DAM
