+++ python/lpf.cpp
@@ -6,7 +6,7 @@
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
 
-#include "sobec/lpf.hpp"
+#include "sobec/lowpassfilter/lpf.hpp"
 
 #include <pinocchio/fwd.hpp>  // to avoid compilation error (https://github.com/loco-3d/crocoddyl/issues/205)
